+++ internal/provider/data_source_fmc_single_sign_on_server.go
@@ -1,0 +1,214 @@
+// Copyright © 2023 Cisco Systems, Inc. and its affiliates.
+// All rights reserved.
+//
+// Licensed under the Mozilla Public License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://mozilla.org/MPL/2.0/
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// SPDX-License-Identifier: MPL-2.0
+
+package provider
+
+// Section below is generated&owned by "gen/generator.go". //template:begin imports
+import (
+	"context"
+	"fmt"
+	"net/url"
+
+	"github.com/CiscoDevNet/terraform-provider-fmc/internal/provider/helpers"
+	"github.com/hashicorp/terraform-plugin-framework-validators/datasourcevalidator"
+	"github.com/hashicorp/terraform-plugin-framework/datasource"
+	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
+	"github.com/hashicorp/terraform-plugin-framework/path"
+	"github.com/hashicorp/terraform-plugin-framework/types"
+	"github.com/hashicorp/terraform-plugin-log/tflog"
+	"github.com/netascode/go-fmc"
+	"github.com/tidwall/gjson"
+)
+
+// End of section. //template:end imports
+
+// Section below is generated&owned by "gen/generator.go". //template:begin model
+
+// Ensure the implementation satisfies the expected interfaces.
+var (
+	_ datasource.DataSource              = &SingleSignOnServerDataSource{}
+	_ datasource.DataSourceWithConfigure = &SingleSignOnServerDataSource{}
+)
+
+func NewSingleSignOnServerDataSource() datasource.DataSource {
+	return &SingleSignOnServerDataSource{}
+}
+
+type SingleSignOnServerDataSource struct {
+	client *fmc.Client
+}
+
+func (d *SingleSignOnServerDataSource) Metadata(_ context.Context, req datasource.MetadataRequest, resp *datasource.MetadataResponse) {
+	resp.TypeName = req.ProviderTypeName + "_single_sign_on_server"
+}
+
+func (d *SingleSignOnServerDataSource) Schema(ctx context.Context, req datasource.SchemaRequest, resp *datasource.SchemaResponse) {
+	resp.Schema = schema.Schema{
+		// This description is used by the documentation generator and the language server.
+		MarkdownDescription: helpers.NewAttributeDescription("This data source reads the Single Sign On Server.").String,
+
+		Attributes: map[string]schema.Attribute{
+			"id": schema.StringAttribute{
+				MarkdownDescription: "Id of the object",
+				Optional:            true,
+				Computed:            true,
+			},
+			"domain": schema.StringAttribute{
+				MarkdownDescription: "Name of the FMC domain",
+				Optional:            true,
+			},
+			"name": schema.StringAttribute{
+				MarkdownDescription: "Name of the Single Sign-On Server object.",
+				Optional:            true,
+				Computed:            true,
+			},
+			"type": schema.StringAttribute{
+				MarkdownDescription: "Type of the object; this value is always 'SSOServer'.",
+				Computed:            true,
+			},
+			"identity_provider_entity_id_url": schema.StringAttribute{
+				MarkdownDescription: "URL that is defined in SAML Identity Provider (IdP) to identify a service provider uniquely.",
+				Computed:            true,
+			},
+			"sso_url": schema.StringAttribute{
+				MarkdownDescription: "URL for signing into the SAML Identity Provider (IdP) server.",
+				Computed:            true,
+			},
+			"logout_url": schema.StringAttribute{
+				MarkdownDescription: "URL for signing out of the SAML Identity Provider (IdP) server.",
+				Computed:            true,
+			},
+			"base_url": schema.StringAttribute{
+				MarkdownDescription: "URL that will redirect the user back to FTD once the Identity Provider (IdP) authentication is done.",
+				Computed:            true,
+			},
+			"identity_provider_certificate_id": schema.StringAttribute{
+				MarkdownDescription: "Id of the Identity Provider (IdP) certificate utilized to verify messages signed by the IdP.",
+				Computed:            true,
+			},
+			"identity_provider_certificate_name": schema.StringAttribute{
+				MarkdownDescription: "Name of the Identity Provider (IdP) certificate utilized to verify messages signed by the IdP.",
+				Computed:            true,
+			},
+			"service_provider_certificate_id": schema.StringAttribute{
+				MarkdownDescription: "Id of the certificate, which will be used to sign the requests and build circle of trust with IdP.",
+				Computed:            true,
+			},
+			"request_signature_type": schema.StringAttribute{
+				MarkdownDescription: "Encryption algorithm to sign the SAML single sign-on requests.",
+				Computed:            true,
+			},
+			"request_timeout": schema.Int64Attribute{
+				MarkdownDescription: "Specify the SAML assertion validity duration (in seconds) for the users to complete the single sign-on request.",
+				Computed:            true,
+			},
+			"identity_provider_accessible_only_on_internal_network": schema.BoolAttribute{
+				MarkdownDescription: "SAML Identity Provider (IdP) resides on the internal network.",
+				Computed:            true,
+			},
+			"request_identity_provider_reauthentication_at_each_login": schema.BoolAttribute{
+				MarkdownDescription: "Authenticate user at each login even if the previous Identity Provider (IdP) session is valid.",
+				Computed:            true,
+			},
+		},
+	}
+}
+func (d *SingleSignOnServerDataSource) ConfigValidators(ctx context.Context) []datasource.ConfigValidator {
+	return []datasource.ConfigValidator{
+		datasourcevalidator.ExactlyOneOf(
+			path.MatchRoot("id"),
+			path.MatchRoot("name"),
+		),
+	}
+}
+
+func (d *SingleSignOnServerDataSource) Configure(_ context.Context, req datasource.ConfigureRequest, _ *datasource.ConfigureResponse) {
+	if req.ProviderData == nil {
+		return
+	}
+
+	d.client = req.ProviderData.(*FmcProviderData).Client
+}
+
+// End of section. //template:end model
+
+// Section below is generated&owned by "gen/generator.go". //template:begin read
+
+func (d *SingleSignOnServerDataSource) Read(ctx context.Context, req datasource.ReadRequest, resp *datasource.ReadResponse) {
+	var config SingleSignOnServer
+
+	// Read config
+	diags := req.Config.Get(ctx, &config)
+	resp.Diagnostics.Append(diags...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	// Set request domain if provided
+	reqMods := [](func(*fmc.Req)){}
+	if !config.Domain.IsNull() && config.Domain.ValueString() != "" {
+		reqMods = append(reqMods, fmc.DomainName(config.Domain.ValueString()))
+	}
+
+	tflog.Debug(ctx, fmt.Sprintf("%s: Beginning Read", config.Id.String()))
+	if config.Id.IsNull() && !config.Name.IsNull() {
+		offset := 0
+		limit := 1000
+		for page := 1; ; page++ {
+			queryString := fmt.Sprintf("?limit=%d&offset=%d&expanded=true", limit, offset)
+			res, err := d.client.Get(config.getPath()+queryString, reqMods...)
+			if err != nil {
+				resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to retrieve objects, got error: %s", err))
+				return
+			}
+			if value := res.Get("items"); len(value.Array()) > 0 {
+				value.ForEach(func(k, v gjson.Result) bool {
+					if config.Name.ValueString() == v.Get("name").String() {
+						config.Id = types.StringValue(v.Get("id").String())
+						tflog.Debug(ctx, fmt.Sprintf("%s: Found object with name '%v', id: %v", config.Id.ValueString(), config.Name.ValueString(), config.Id.ValueString()))
+						return false
+					}
+					return true
+				})
+			}
+			if !config.Id.IsNull() || !res.Get("paging.next.0").Exists() {
+				break
+			}
+			offset += limit
+		}
+
+		if config.Id.IsNull() {
+			resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to find object with name: %v", config.Name.ValueString()))
+			return
+		}
+	}
+	urlPath := config.getPath() + "/" + url.QueryEscape(config.Id.ValueString())
+	res, err := d.client.Get(urlPath, reqMods...)
+	if err != nil {
+		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to retrieve object, got error: %s", err))
+		return
+	}
+
+	config.fromBody(ctx, res)
+
+	tflog.Debug(ctx, fmt.Sprintf("%s: Read finished successfully", config.Id.ValueString()))
+
+	diags = resp.State.Set(ctx, &config)
+	resp.Diagnostics.Append(diags...)
+}
+
+// End of section. //template:end read